requirements.txt
asgiref==3.8.1
bigcode==1.0.9
certifi==2024.7.4
Django==5.0.7
JazzminTM==1.0.6.1
pillow==10.4.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
